Implement real-time status icon updates
Woah, that was a wild ride, but not because of the Socket.IO
implementation (this was rather straightforward in this case), but
due to numerous bugs related to changing status icons to available and
unavailable when going online and offline respectively. This turned out
to be rather difficult to manage consistently, but I think I've got it
under control now.