Turn emoticons into a single object
I realised once again that it is much better to create a single object
with objects as values instead of creating an array of objects. If
iteration is your concern for the former, worry no more, you can easily
turn the object into the array of objects with Object.values().