Add automatic emoticon insertion while typing
It only works if the caret is at the end of the input field, but it
alone is good enough. I tried my best to implement it for the entire
input field text, but storing and restoring caret position after
modifying the input field innerHTML value containing additional nodes
proved way, way too much for me to handle.