Rework automatic status icon changing
It is now handled on the backend, which seems to be much more reliable
and deals with disconnections gracefully, making sure DB data are
always accurate. By getting rid of lots of conflicting client-side
code, I also fixed a bug where the status icon set manually by the user
was reset to available when the user came back to the view with tabs
from chats.